+++ Sources/Lux/InjectorImplementations/Json/JSONDelegate.swift
@@ -4,6 +4,7 @@ extension JSONDelegate {
     public static func theme(_ theme: ColorTheme) -> JSONDelegate {
         switch theme {
         case .dracula: return .dracula
+        case .xcodeLight: return .xcodeLight
         }
     }
 }