+++ Sources/Lux/InjectorImplementations/Plist/PlistDelegate.swift
@@ -4,6 +4,7 @@ extension PlistDelegate {
     public static func theme(_ theme: ColorTheme) -> PlistDelegate {
         switch theme {
         case .dracula: return .dracula
+        case .xcodeLight: return .xcodeLight
         }
     }
 }