+++ Sources/Lux/InjectorImplementations/Swift/SwiftCategory.swift
@@ -1,6 +1,6 @@
 import Foundation
 import Splash
-#if !os(macOS)
+#if os(iOS)
 import UIKit
 #else
 import AppKit