+++ Sources/Lux/InjectorImplementations/Xml/XMLDelegate.swift
@@ -4,6 +4,7 @@ extension XMLDelegate {
     public static func theme(_ theme: ColorTheme) -> XMLDelegate {
         switch theme {
         case .dracula: return .dracula
+        case .xcodeLight: return .xcodeLight
         }
     }
 }