+++ Sources/Lux/Themes/Xcode Light/JSONXcodeDefaultDelegate.swift
@@ -1,0 +1,28 @@
+#if os(iOS)
+import UIKit
+#elseif os(macOS)
+import AppKit
+#endif
+
+public extension JSONDelegate {
+    static var xcodeLight: JSONDelegate { XcodeLightDelegate() }
+}
+
+final class XcodeLightDelegate: JSONDelegate, XcodeLightThemeInjectorDelegate {
+
+    override func color(for category: JSONCategory) -> Color {
+        switch category {
+        case .keyName: return Color.xcodeLight.property
+        case .keyValue: return Color.xcodeLight.plainText
+        case .punctuation: return Color.xcodeLight.comment
+        }
+    }
+
+    override func terminalModifier(for category: JSONCategory) -> TerminalModifier {
+        switch category {
+        case .keyName: return TerminalModifier.xcodeLight.property
+        case .keyValue: return TerminalModifier.xcodeLight.plainText
+        case .punctuation: return TerminalModifier.xcodeLight.comment
+        }
+    }
+}